refactor: share recorder take model
