+++ src/lib/recorder/take-comp.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from "vitest";
 import { renderTakeComp } from "./take-comp.ts";
 import { deriveTakeRegions } from "./take-regions.ts";
+import type { TakeState } from "./take.ts";
 
 describe(renderTakeComp, () => {
   it("renders newer overlapping samples over an older take", () => {
@@ -8,17 +9,19 @@ describe(renderTakeComp, () => {
     const takes = [
       {
         id: "old",
+        number: 1,
         timelineOffset: 0,
         duration: 2,
         buffer: createBuffer({ sampleRate: 4, samples: Array(8).fill(1) }),
       },
       {
         id: "new",
+        number: 2,
         timelineOffset: 0.75,
         duration: 0.5,
         buffer: createBuffer({ sampleRate: 4, samples: [2, 2] }),
       },
-    ];
+    ] satisfies TakeState[];
     const result = renderTakeComp({
       context,
       regions: deriveTakeRegions(takes),
@@ -34,11 +37,12 @@ describe(renderTakeComp, () => {
     const takes = [
       {
         id: "take",
+        number: 1,
         timelineOffset: 0,
         duration: 1,
         buffer: createBuffer({ sampleRate: 2, samples: [0, 1] }),
       },
-    ];
+    ] satisfies TakeState[];
     const result = renderTakeComp({
       context: createContext(4),
       regions: deriveTakeRegions(takes),
@@ -52,11 +56,12 @@ describe(renderTakeComp, () => {
     const takes = [
       {
         id: "take",
+        number: 1,
         timelineOffset: 1,
         duration: 1,
         buffer: createBuffer({ sampleRate: 2, samples: [1, 1] }),
       },
-    ];
+    ] satisfies TakeState[];
     const result = renderTakeComp({
       context: createContext(2),
       regions: deriveTakeRegions(takes),
